pyproject.toml
[project]
name = "pubmed-navigator-workshop"
version = "0.1.0"
description = ""
requires-python = ">=3.13"

dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workshop"]
